Fixed a bug with how I was sending plugin messages to the proxy.
Different classes implement the "sendPluginMessage" differently. For the Server object, it does a "for (Player player : getOnlinePlayers())" call, which means each player online is sending the message, regardless of who ran the command. Not ideal.

Had to fix that as well as some logic in the Velocity side. Now I gotta re-record the whole tutorial. Oh well